Added linking new weapons stats to warlordsplayer in pve

+++ src/main/java/com/ebicep/warlords/pve/weapons/weapontypes/StarPieceBonus.java
@@ -18,8 +18,8 @@ public interface StarPieceBonus {
 
     int getStarPieceBonusValue();
 
-    default double getStarPieceBonusMultiplicativeValue() {
-        return 1 + getStarPieceBonusValue() / 100.0;
+    default float getStarPieceBonusMultiplicativeValue() {
+        return 1 + getStarPieceBonusValue() / 100.0f;
     }
 
     default List<WeaponStats> getRandomStatBonus() {
